+++ vortex-scan/Cargo.toml
@@ -38,7 +38,6 @@ tokio = { workspace = true, features = ["rt-multi-thread"], optional = true }
 
 [dev-dependencies]
 vortex-layout = { workspace = true, features = ["test-harness"] }
-vortex-scan = { path = ".", features = ["test-harness"] }
 
 [target.'cfg(loom)'.dev-dependencies]
 loom = { workspace = true }
@@ -46,7 +45,6 @@ loom = { workspace = true }
 [features]
 default = []
 roaring = ["dep:roaring", "vortex-layout/roaring"]
-test-harness = []
 tokio = ["dep:tokio"]
 
 [lints]
